fix: migrate landing page to server component for better perf, add studio route logic, and unify button styles
