+++ src/app/page.tsx
@@ -1,24 +1,19 @@
-'use client'
-
 import { Icons } from '@/components/icons'
-import DuolingoButton, {
-  baseStyles,
-  sizeStyles,
-  variantStyles,
-} from '@/components/ui/duolingo-button'
 import GitHubStarButton from '@/components/ui/github-star-button'
 import { cn } from '@/lib/utils'
 import MuxPlayer from '@mux/mux-player-react'
 import { Menu } from 'lucide-react'
-import dynamic from 'next/dynamic'
 import Link from 'next/link'
+import { auth } from '@/lib/auth'
+import { headers } from 'next/headers'
+import TestimonialsComponent from '@/components/testimonials-component'
+import DuolingoButton from '@/components/ui/duolingo-button'
+import { baseStyles, sizeStyles, variantStyles } from '@/components/ui/duolingo-styles'
 
-const Testimonials = dynamic(
-  () => import('@/app/testimonials').then((mod) => ({ default: mod.Testimonials })),
-  { ssr: false },
-)
-
-const Page = () => {
+const Page = async () => {
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  })
   return (
     <>
       <section className="bg-gray-100">
@@ -49,12 +44,21 @@ const Page = () => {
                     className="whitespace-nowrap"
                     repo="joschan21/contentport"
                   />
-                  <Link
-                    className={cn(baseStyles, variantStyles.primary, sizeStyles.sm)}
-                    href="/login"
-                  >
-                    Get Started
-                  </Link>
+                  {session?.user ? (
+                    <Link
+                      href="/studio"
+                      className={cn(baseStyles, variantStyles.primary, sizeStyles.sm)}
+                    >
+                      Studio
+                    </Link>
+                  ) : (
+                    <Link
+                      className={cn(baseStyles, variantStyles.primary, sizeStyles.sm)}
+                      href="/login"
+                    >
+                      Get Started
+                    </Link>
+                  )}
                 </div>
               </div>
             </nav>
@@ -79,7 +83,7 @@ const Page = () => {
               <div className="max-w-4xl mx-auto text-center">
                 <div className="flex flex-col justify-center items-center">
                   <div className="flex items-center justify-center gap-2 text-sm/6 mb-4 py-1 px-4 rounded-full bg-gray-200/50 text-gray-800">
-                  🎉 now open to everyone! 🎉
+                    🎉 now open to everyone! 🎉
                   </div>
                   <h1 className="text-5xl font-semibold tracking-tight text-balance text-gray-900 sm:text-6xl">
                     The <span className="text-indigo-600">content engine </span> for
@@ -118,11 +122,19 @@ const Page = () => {
                     </div>
 
                     <div className="flex mt-4 flex-col gap-2 max-w-sm w-full">
-                      <Link href="/login">
-                        <DuolingoButton className="w-full h-12 sm:px-8">
-                          Start Posting More →
-                        </DuolingoButton>
-                      </Link>
+                      {session?.user ? (
+                        <Link href="/studio">
+                          <DuolingoButton className="w-full h-12 sm:px-8">
+                            Start Posting More →
+                          </DuolingoButton>
+                        </Link>
+                      ) : (
+                        <Link href="/login">
+                          <DuolingoButton className="w-full h-12 sm:px-8">
+                            Start Posting More →
+                          </DuolingoButton>
+                        </Link>
+                      )}
                     </div>
 
                     <div className="mt-2 flex items-center justify-center gap-4">
@@ -220,7 +232,7 @@ const Page = () => {
               />
             </div>
 
-            <Testimonials />
+            <TestimonialsComponent />
           </div>
         </div>
       </section>
